meet/requirements.txt
Django==2.2.19
django-import-export-pandas==0.1.1
pytz==2022.1
sqlparse==0.4.2
